refactor(broker): simplify ActorFutureAggregator
